requirements.txt
astroid==2.8.4
base58==2.1.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
names==0.3.0
p2pnetwork==1.1
platformdirs==2.4.0
pylint==2.11.1
pymerkle==2.0.2
python-gnupg==0.4.7
PyYAML==6.0
toml==0.10.2
tqdm==4.62.3
typed-ast==1.4.3
typing-extensions==3.10.0.2
wrapt==1.13.3
